:recycle: refactor chatbot service
